sox: formats: implement musl pipe rewinding
The big and only reason why pipe rewinding wasn't implemented for musl
is that musl upstream is very vehement about not exposing the struct
FILE definition in contrast to GNU & the rest of libc's which share it.

So the only way to rewind the pipe in musl is to copy the actual
opaque struct definition over, then it works great.

Of course this may break in the future if the musl FILE implementation
changes but that is very rare and it's better than silently disabling
the feature.

Upstream-Status: Innappropriate [OE/Musl specific]

Signed-off-by: Ioan-Adrian Ratiu <adi@adirat.com>